requirements.txt
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.17.1
fonttools==4.40.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.3
ipython==8.14.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook_shim==0.2.3
numpy==1.25.0
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.1
rpds-py==0.9.2
scikit-learn==1.2.2
scipy==1.11.0
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
yarg==0.1.9
zipp==3.15.0
